requirements.txt
PyPDF2
pdfplumber
python-docx
scikit-learn
sklearn
pandas
matplotlib
